Refactor some shared code
