+++ src/Wintel64/common/dirlist.h
@@ -1,0 +1,9 @@
+#ifndef DIRLIST_H
+#define DIRLIST_H
+
+#include <vector>
+#include <string>
+
+std::vector<std::string> dir_to_list(const std::string &dirname);
+
+#endif
